refactor: map 전역과 겹치지 않게 onboardingmap으로 파일명 변경
